backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "arize-phoenix>=8.26.3",
    "baml-py>=0.85.0",
    "fastapi[standard]>=0.115.12",
    "gitpython>=3.1.44",
    "llama-index>=0.12.32",
    "llama-index-utils-workflow>=0.3.1",
    "marimo>=0.13.1",
    "matplotlib>=3.10.3",
    "networkx>=3.4.2",
    "openinference-instrumentation-llama-index>=4.2.1",
    "opentelemetry-exporter-otlp>=1.32.1",
    "psycopg[binary]>=3.2.6",
    "pydantic-ai>=0.4.6",
    "pydantic-ai-slim[ag-ui]>=0.4.6",
    "python-dotenv>=1.1.0",
    "sqlmodel>=0.0.24",
    "streamlit>=1.45.1",
    "supabase>=2.15.1",
    "uvicorn>=0.34.2",
    "watchdog>=6.0.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["app", "lib", "config", "storage", "eval"]

[tool.uv.sources]
app = { path = "./app" }
